feat: add CreatePayment command

+++ internal/domain/payment.go
@@ -3,10 +3,10 @@ package domain
 import "time"
 
 type Payment struct {
-	Id          int
+	Id          int64
 	Date        time.Time
 	Description string
 	Category    Category
 	Value       float32
-	UserId      int
+	UserId      int64
 }
